+++ src/Utilities.jl
@@ -15,3 +15,22 @@ function check_exists(path::String)
     throw(ErrorException(string("Necessary path doesn't exist on system: ", path)))
   end
 end
+
+"""
+    sym_or_cp(path_orig::String, path_dest::String, sym::Bool)
+
+Helper function that if sym is true, symlink the original path to the destination
+if sym is false, cp the original path to the destination
+"""
+function sym_or_cp(path_orig::String, path_dest::String, sym::Bool, upgrade::Bool)
+  if upgrade
+    rm(path_dest; force=true)
+  end
+  if !ispath(path_dest)
+    if sym
+      symlink(path_orig, path_dest)
+    else
+      cp(path_orig, path_dest; follow_symlinks=true)
+    end
+  end
+end