+++ CHANGELOG.md
@@ -7,11 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.1.0]
+
 ### Added
 
+- Added custom prompts
+- Added Windows command prompt support
 - Added usage display if `help` option provided or no environment directories
   provided
 - Added `upgrade` option to upgrade julia in place if exists before creating
 - Added `clear` option to clear environment directory if exists before creating
 - Added ability to create multiple environments at once
 - `venv` function to create self-contained julia environments
+
+[Unreleased]: https://github.com/mehalter/VirtualEnv.jl/compare/v0.1.0...HEAD
+[v0.1.0]: https://github.com/mehalter/VirtualEnv.jl/releases/tag/v0.0.1